Second step
Removed the hard coded parts. There is a "company" model and a "options" model. Company model decides what company we're predicting, options model decides the start and end dates of the prediction, and the training data.

You can display the appropriate prediction by going to /display/*companyID*/*optionsID*

Currently there is only one company and option (facebook, predict from 2021 to 2022, train on data from 2020 to 2021) but we can add more